+++ src/main/java/com/counter/budget/budgetcounterbe/dto/TransactionResponse.java
@@ -2,6 +2,7 @@ package com.counter.budget.budgetcounterbe.dto;
 
 import com.counter.budget.budgetcounterbe.model.TransactionType;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -10,5 +11,5 @@ public record TransactionResponse(
         String timestamp,
         List<UUID> bucketTransactions,
         TransactionType type,
-        float amount
+        BigDecimal amount
 ) {}
